only download words that aren't present, in case google decides to rate-limit

+++ Python/WordTraderRadio/Song.py
@@ -191,19 +191,20 @@ class Song:
             os.makedirs(self.WAVS_DIR)
 
         for s in sylHash:
-            response = urllib2.urlopen(urllib2.Request(url+urllib.quote(s), None, header))
-            responseBytes = response.read()
             mp3FilePath = self.MP3S_DIR+s.decode('iso-8859-1')+'.mp3'
             wavFilePath = mp3FilePath.replace('mp3','wav')
-            f = open(mp3FilePath, 'wb')
-            f.write(responseBytes)
-            f.close()
-            ffParams = "-y -i %s -ar 44100 %s"%(escSpace(mp3FilePath), escSpace(wavFilePath))
-            subprocess.call('ffmpeg '+ffParams, shell=True, stdout=self.FNULL, stderr=subprocess.STDOUT)
-            wavWave = wave.open(wavFilePath)
-            wavLength = wavWave.getnframes()/float(wavWave.getframerate())
-            wavWave.close()
-            sylHash[s] = (wavFilePath, wavLength)
+            if not os.path.isfile(mp3FilePath):
+                response = urllib2.urlopen(urllib2.Request(url+urllib.quote(s), None, header))
+                responseBytes = response.read()
+                f = open(mp3FilePath, 'wb')
+                f.write(responseBytes)
+                f.close()
+                ffParams = "-y -i %s -ar 44100 %s"%(escSpace(mp3FilePath), escSpace(wavFilePath))
+                subprocess.call('ffmpeg '+ffParams, shell=True, stdout=self.FNULL, stderr=subprocess.STDOUT)
+                wavWave = wave.open(wavFilePath)
+                wavLength = wavWave.getnframes()/float(wavWave.getframerate())
+                wavWave.close()
+                sylHash[s] = (wavFilePath, wavLength)
         shutil.rmtree(self.MP3S_DIR)
 
         for (i, (s,t,p,d)) in enumerate(self.tonedSyls):
